+++ lib/flutter_native_text_input.dart
@@ -1,6 +1,10 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 
 enum ReturnKeyType {
   defaultAction,
@@ -80,24 +84,26 @@ enum KeyboardType {
 
   /// A default keyboard type with URL-oriented addition (shows space . prominently).
   webSearch,
+
   // A number pad (0-9) that will always be ASCII digits. Falls back to KeyboardType.numberPad below iOS 10.
   asciiCapableNumberPad
 }
 
 class NativeTextInput extends StatefulWidget {
+  static const viewType = 'flutter_native_text_input';
+
   const NativeTextInput({
     Key? key,
     this.controller,
-    this.cursorColor,
     this.decoration,
     this.focusNode,
-    this.keyboardAppearance,
+    this.iosOptions,
     this.keyboardType = KeyboardType.defaultType,
     this.maxLines = 1,
     this.minHeightPadding = 18,
     this.minLines = 1,
     this.placeholder,
-    this.placeholderStyle,
+    this.placeholderColor,
     this.returnKeyType = ReturnKeyType.done,
     this.style,
     this.textAlign = TextAlign.start,
@@ -114,12 +120,6 @@ class NativeTextInput extends StatefulWidget {
   /// Default: null, this widget will create its own [TextEditingController].
   final TextEditingController? controller;
 
-  /// The color of the cursor
-  /// (https://api.flutter.dev/flutter/material/TextField/cursorColor.html)
-  ///
-  /// Default: null
-  final Color? cursorColor;
-
   /// Controls the BoxDecoration of the box behind the text input
   /// (https://api.flutter.dev/flutter/cupertino/CupertinoTextField/decoration.html)
   ///
@@ -132,11 +132,10 @@ class NativeTextInput extends StatefulWidget {
   /// Default: null
   final FocusNode? focusNode;
 
-  /// The appearance of the keyboard
-  /// (https://api.flutter.dev/flutter/material/TextField/keyboardAppearance.html)
+  /// iOS only options (cursorColor, keyboardAppearance)
   ///
   /// Default: null
-  final Brightness? keyboardAppearance;
+  final IosOptions? iosOptions;
 
   /// Type of keyboard to display for a given text-based view
   /// (https://developer.apple.com/documentation/uikit/uikeyboardtype)
@@ -167,11 +166,10 @@ class NativeTextInput extends StatefulWidget {
   /// Default: null
   final String? placeholder;
 
-  /// The style to use for the placeholder text. [Only `fontSize`, `fontWeight`, `color` are supported]
-  /// (https://api.flutter.dev/flutter/cupertino/CupertinoTextField/placeholderStyle.html)
+  /// The text color to use for the placeholder text
   ///
   /// Default: null
-  final TextStyle? placeholderStyle;
+  final Color? placeholderColor;
 
   /// Constants that specify the text string that displays in the Return key of a keyboard
   /// (https://developer.apple.com/documentation/uikit/uireturnkeytype)
@@ -225,6 +223,22 @@ class NativeTextInput extends StatefulWidget {
   State<StatefulWidget> createState() => _NativeTextInputState();
 }
 
+class IosOptions {
+  /// The color of the cursor
+  /// (https://api.flutter.dev/flutter/material/TextField/cursorColor.html)
+  ///
+  /// Default: null
+  final Color? cursorColor;
+
+  /// The appearance of the keyboard
+  /// (https://api.flutter.dev/flutter/material/TextField/keyboardAppearance.html)
+  ///
+  /// Default: null
+  final Brightness? keyboardAppearance;
+
+  IosOptions({this.cursorColor, this.keyboardAppearance});
+}
+
 class _NativeTextInputState extends State<NativeTextInput> {
   late MethodChannel _channel;
 
@@ -262,6 +276,59 @@ class _NativeTextInputState extends State<NativeTextInput> {
     }
   }
 
+  Widget _platformView(BoxConstraints layout) {
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return PlatformViewLink(
+          viewType: NativeTextInput.viewType,
+          surfaceFactory: (context, controller) => AndroidViewSurface(
+            controller: controller as AndroidViewController,
+            hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+            gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
+          ),
+          onCreatePlatformView: (PlatformViewCreationParams params) {
+            return PlatformViewsService.initSurfaceAndroidView(
+              id: params.id,
+              viewType: NativeTextInput.viewType,
+              layoutDirection: TextDirection.ltr,
+              creationParams: _buildCreationParams(layout),
+              creationParamsCodec: const StandardMessageCodec(),
+            )
+              ..addOnPlatformViewCreatedListener((_) {
+                params.onPlatformViewCreated(_);
+                _createMethodChannel(_);
+              })
+              ..create();
+          },
+        );
+      case TargetPlatform.iOS:
+        return UiKitView(
+          viewType: NativeTextInput.viewType,
+          creationParamsCodec: const StandardMessageCodec(),
+          creationParams: _buildCreationParams(layout),
+          onPlatformViewCreated: _createMethodChannel,
+        );
+      default:
+        return CupertinoTextField(
+          controller: widget.controller,
+          cursorColor: widget.iosOptions?.cursorColor,
+          decoration: BoxDecoration(
+            border: Border.all(width: 0, color: Colors.transparent),
+          ),
+          focusNode: widget.focusNode,
+          keyboardAppearance: widget.iosOptions?.keyboardAppearance,
+          maxLines: widget.maxLines,
+          minLines: widget.minLines,
+          placeholder: widget.placeholder,
+          placeholderStyle: TextStyle(color: widget.placeholderColor),
+          textAlign: widget.textAlign,
+          textCapitalization: widget.textCapitalization,
+          onChanged: widget.onChanged,
+          onSubmitted: widget.onSubmitted,
+        );
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return ConstrainedBox(
@@ -272,12 +339,7 @@ class _NativeTextInputState extends State<NativeTextInput> {
       child: LayoutBuilder(
         builder: (context, layout) => Container(
           decoration: widget.decoration,
-          child: UiKitView(
-            viewType: "flutter_native_text_input",
-            creationParamsCodec: const StandardMessageCodec(),
-            creationParams: _buildCreationParams(layout),
-            onPlatformViewCreated: _createMethodChannel,
-          ),
+          child: _platformView(layout),
         ),
       ),
     );
@@ -304,13 +366,14 @@ class _NativeTextInputState extends State<NativeTextInput> {
     Map<String, dynamic> params = {
       "maxLines": widget.maxLines,
       "minHeightPadding": widget.minHeightPadding,
+      "minLines": widget.minLines,
       "placeholder": widget.placeholder ?? "",
       "returnKeyType": widget.returnKeyType.toString(),
       "text": _effectiveController.text,
       "textAlign": widget.textAlign.toString(),
       "textCapitalization": widget.textCapitalization.toString(),
       "textContentType": widget.textContentType?.toString(),
-      "keyboardAppearance": widget.keyboardAppearance.toString(),
+      "keyboardAppearance": widget.iosOptions?.keyboardAppearance.toString(),
       "keyboardType": widget.keyboardType.toString(),
       "width": constraints.maxWidth,
     };
@@ -341,43 +404,26 @@ class _NativeTextInputState extends State<NativeTextInput> {
       };
     }
 
-    if (widget.placeholderStyle != null &&
-        widget.placeholderStyle?.fontSize != null) {
-      params = {
-        ...params,
-        "placeholderFontSize": widget.placeholderStyle?.fontSize,
-      };
-    }
-
-    if (widget.placeholderStyle != null &&
-        widget.placeholderStyle?.fontWeight != null) {
-      params = {
-        ...params,
-        "placeholderFontWeight": widget.placeholderStyle?.fontWeight.toString(),
-      };
-    }
-
-    if (widget.placeholderStyle != null &&
-        widget.placeholderStyle?.color != null) {
+    if (widget.placeholderColor != null) {
       params = {
         ...params,
         "placeholderFontColor": {
-          "red": widget.placeholderStyle?.color?.red,
-          "green": widget.placeholderStyle?.color?.green,
-          "blue": widget.placeholderStyle?.color?.blue,
-          "alpha": widget.placeholderStyle?.color?.alpha,
+          "red": widget.placeholderColor?.red,
+          "green": widget.placeholderColor?.green,
+          "blue": widget.placeholderColor?.blue,
+          "alpha": widget.placeholderColor?.alpha,
         },
       };
     }
 
-    if (widget.cursorColor != null) {
+    if (widget.iosOptions?.cursorColor != null) {
       params = {
         ...params,
         "cursorColor": {
-          "red": widget.cursorColor?.red,
-          "green": widget.cursorColor?.green,
-          "blue": widget.cursorColor?.blue,
-          "alpha": widget.cursorColor?.alpha,
+          "red": widget.iosOptions!.cursorColor?.red,
+          "green": widget.iosOptions!.cursorColor?.green,
+          "blue": widget.iosOptions!.cursorColor?.blue,
+          "alpha": widget.iosOptions!.cursorColor?.alpha,
         },
       };
     }
@@ -414,6 +460,9 @@ class _NativeTextInputState extends State<NativeTextInput> {
     if (!_isMultiline) return _minHeight;
     if (_contentHeight > _minHeight && widget.maxLines > 0) {
       double maxLineHeight = widget.maxLines * _lineHeight;
+      if (defaultTargetPlatform == TargetPlatform.android) {
+        maxLineHeight += widget.minHeightPadding;
+      }
       return _contentHeight > maxLineHeight ? maxLineHeight : _contentHeight;
     }
     if (_contentHeight > _minHeight) return _contentHeight;
@@ -423,8 +472,9 @@ class _NativeTextInputState extends State<NativeTextInput> {
   // input control methods
   void _inputStarted() {
     _scrollIntoView();
-    if (!_effectiveFocusNode.hasFocus)
+    if (!_effectiveFocusNode.hasFocus) {
       FocusScope.of(context).requestFocus(_effectiveFocusNode);
+    }
   }
 
   void _inputFinished(String? text) {
@@ -435,7 +485,7 @@ class _NativeTextInputState extends State<NativeTextInput> {
       if (_effectiveFocusNode.hasFocus) FocusScope.of(context).unfocus();
     }
     if (widget.onSubmitted != null) {
-      Future.delayed(Duration(milliseconds: 100), () {
+      Future.delayed(const Duration(milliseconds: 100), () {
         widget.onSubmitted!(text);
       });
     }